// laptopstore/src/app/dto/product-cart-dto.ts
export interface ProductCartDto {
   id?: string;
   ordersId?: string;
   name?: string;
   urlImgs?: string;
   price?: number;
   color?: string;
   discountName?: string;
   discountVal?: number;
   endDay?: string;
   quality?: string;
   ram?: string;
   cpu?: string;
   memory?: string;
   memtype?: string;
   ssdVal?: string;
   weight?: number;
   ordersDay?: string;
   productGroupId?: number;
   locationShip?: string;
}